Arreglos de detalles y objetos

+++ Demo01.api/DTO/ProductsDTO/ProductDTO.cs
@@ -1,4 +1,4 @@
 ﻿namespace Demo01.api.DTO.ProductsDTO
 {
-    public record ProductDTO(string productName, float unitPrice, int quantity);
+    public record ProductDTO(string productName, decimal unitPrice, Int16 quantity);
 }
