+++ Demo01.api/Domain/Entities/Products.cs
@@ -5,9 +5,9 @@ namespace Demo01.api.Domain.Entities
     public class Products
     {
         public string ProductName { get; set; }
-        public float UnitPrice {  get; set; }
-        public int Quantity { get; set; }
-       public Products(string productName, float unitPrice, int quantity) {
+        public decimal UnitPrice {  get; set; }
+        public Int16 Quantity { get; set; }
+       public Products(string productName, decimal unitPrice, Int16 quantity) {
             ProductName = productName;
             UnitPrice = unitPrice;
             Quantity = quantity;
